Rename directory "plugins" to "plugin"
Fixes #152
