+++ plugin/grpc/grpc_test.go
@@ -20,7 +20,7 @@ import (
 
 	"golang.org/x/net/context"
 
-	"go.opencensus.io/plugins/grpc/grpcstats"
+	"go.opencensus.io/plugin/grpc/grpcstats"
 	"go.opencensus.io/trace"
 
 	"google.golang.org/grpc/stats"
