+++ plugin/grpc/grpctrace/testdata/generate.sh
