+++ plugin/grpc/grpctrace/testdata/test.proto
